Added definition importing untested

+++ services/Radio-Search.Importer/Radio-Search.Importer.Canada/Radio-Search.Importer.Canada.Data/Models/License/RegulatoryService.cs
@@ -2,6 +2,6 @@
 {
     public class RegulatoryService : MultiLanguageEntry
     {
-        public int RegulatoryServiceID { get; set; }
+        public short RegulatoryServiceID { get; set; }
     }
 }
