+++ services/Radio-Search.Importer/Radio-Search.Importer.Canada/Radio-Search.Importer.Canada.Data/Models/License/ServiceType.cs
@@ -2,6 +2,6 @@
 {
     public class ServiceType : MultiLanguageEntry
     {
-        public int ServiceTypeID { get; set; }
+        public short ServiceTypeID { get; set; }
     }
 }
