+++ services/Radio-Search.Importer/Radio-Search.Importer.Canada/Radio-Search.Importer.Canada.Data/Models/License/StandbyTransmitterInformation.cs
@@ -2,6 +2,6 @@
 {
     public class StandbyTransmitterInformation : MultiLanguageEntry
     {
-        public int StandbyTransmitterInformationID { get; set; }
+        public short StandbyTransmitterInformationID { get; set; }
     }
 }
