+++ services/Radio-Search.Importer/Radio-Search.Importer.Canada/Radio-Search.Importer.Canada.Data/Models/License/SubserviceType.cs
@@ -2,6 +2,6 @@
 {
     public class SubserviceType : MultiLanguageEntry
     {
-        public int SubserviceTypeID { get; set; }
+        public short SubserviceTypeID { get; set; }
     }
 }
